Finished a initial data model for symbol base and price data. Implemented the symbol search view.

+++ finance-tracker/components/PercentagePill.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct PercentagePill: View {
-    let percentage: Float
+    let percentage: Double
     
     var body: some View {
         VStack {
